Create an adapter for TAS
