+++ src/extension.ts
@@ -10,10 +10,9 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { AzureUserInput, callWithTelemetryAndErrorHandling, createAzExtOutputChannel, createTelemetryReporter, IActionContext, registerUIExtensionVariables, UserCancelledError } from 'vscode-azureextensionui';
 import { ConfigurationParams, DidChangeConfigurationNotification, DocumentSelector, LanguageClient, LanguageClientOptions, Middleware, ServerOptions, TransportKind } from 'vscode-languageclient/lib/main';
-import * as tas from 'vscode-tas-client';
 import { registerCommands } from './commands/registerCommands';
 import { LegacyDockerDebugConfigProvider } from './configureWorkspace/LegacyDockerDebugConfigProvider';
-import { COMPOSE_FILE_GLOB_PATTERN, extensionId } from './constants';
+import { COMPOSE_FILE_GLOB_PATTERN } from './constants';
 import { registerDebugConfigurationProvider } from './debugging/coreclr/registerDebugConfigurationProvider';
 import { registerDebugProvider } from './debugging/DebugHelper';
 import { DockerComposeCompletionItemProvider } from './dockerCompose/dockerComposeCompletionItemProvider';
@@ -25,6 +24,7 @@ import { ext } from './extensionVariables';
 import { localize } from './localize';
 import { registerListeners } from './registerListeners';
 import { registerTaskProviders } from './tasks/TaskHelper';
+import { ExperimentationServiceAdapter } from './telemetry/ExperimentationServiceAdapter';
 import { registerActiveUseSurvey } from './telemetry/surveys/activeUseSurvey';
 import { TelemetryPublisher } from './telemetry/TelemetryPublisher';
 import { TelemetryReporterProxy } from './telemetry/TelemetryReporterProxy';
@@ -69,26 +69,8 @@ function initializeExtensionVariables(ctx: vscode.ExtensionContext): void {
     ctx.subscriptions.push(registerActiveUseSurvey(publisher, ctx.globalState));
 
     const telemetryReporterProxy = new TelemetryReporterProxy(publisher, createTelemetryReporter(ctx));
+    ext.experimentationService = new ExperimentationServiceAdapter(ctx.globalState, telemetryReporterProxy);
     ext.reporter = telemetryReporterProxy;
-
-    const extensionVersion = getExtensionVersion() ?? '1';
-    let targetPopulation: tas.TargetPopulation;
-
-    if (ext.runningTests) {
-        targetPopulation = tas.TargetPopulation.Team;
-    } else if (/alpha/ig.test(extensionVersion)) {
-        targetPopulation = tas.TargetPopulation.Insiders;
-    } else {
-        targetPopulation = tas.TargetPopulation.Public;
-    }
-
-    ext.experimentationService = tas.getExperimentationService(
-        extensionId,
-        extensionVersion,
-        targetPopulation,
-        telemetryReporterProxy,
-        ctx.globalState
-    );
 
     if (!ext.keytar) {
         ext.keytar = Keytar.tryCreate();
@@ -184,12 +166,6 @@ async function getDockerInstallationID(): Promise<string> {
     }
 
     return result;
-}
-
-function getExtensionVersion(): string | undefined {
-    const extension = vscode.extensions.getExtension(extensionId);
-    // eslint-disable-next-line @typescript-eslint/tslint/config
-    return extension?.packageJSON?.version;
 }
 
 /**
