+++ src/extensionVariables.ts
@@ -5,9 +5,9 @@
 
 import * as Dockerode from 'dockerode';
 import * as osNode from 'os';
-import { IExperimentationService } from 'tas-client';
 import { ExtensionContext, TreeView } from "vscode";
 import { AzExtTreeDataProvider, AzExtTreeItem, IAzExtOutputChannel, IAzureUserInput, ITelemetryReporter } from "vscode-azureextensionui";
+import { IExperimentationServiceAdapter } from './telemetry/ExperimentationServiceAdapter';
 import { ContainersTreeItem } from './tree/containers/ContainersTreeItem';
 import { ImagesTreeItem } from './tree/images/ImagesTreeItem';
 import { NetworksTreeItem } from './tree/networks/NetworksTreeItem';
@@ -25,7 +25,7 @@ export namespace ext {
     export let outputChannel: IAzExtOutputChannel;
     export let ui: IAzureUserInput;
     export let reporter: ITelemetryReporter;
-    export let experimentationService: IExperimentationService;
+    export let experimentationService: IExperimentationServiceAdapter;
     export let terminalProvider: ITerminalProvider;
     export let keytar: IKeytar | undefined;
     export let dockerode: Dockerode;
